api/requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.9.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.2
requests==2.19.1
sentry-sdk==0.14.4
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
wrapt==1.10.11
zeroc-ice==3.6.5
Pillow==6.2.2
omero-py==5.6.2
